Fix unsigned and cycle errors

+++ models/virtual_currency.go
@@ -2,13 +2,13 @@ package models
 
 
 
-func StartCredits() (uint64){
+func StartCredits() (int64){
 	return 100
 }
 
-func BandwidthToCredits(bandwidthInMegabits uint64) (uint64){
-	return bandwidthInMegabits / 10
+func BandwidthToCredits(bandwidthInMegabits uint64) (int64){
+	return int64(bandwidthInMegabits) / 10
 }
-func CreditsToBandwidthInMegabits(credits uint64)(uint64){
-	return credits * 10
+func CreditsToBandwidthInMegabits(credits uint64)(int64){
+	return int64(credits) * 10
 }
